requirements.txt
streamlit>=1.55.0
openai>=2.28.0
requests>=2.32.0
python-dotenv>=1.2.0
httpx>=0.28.0
urllib3>=2.6.0
